Last commit broke div waiting/added error handling
waitToDisappear wouldn't check a second time if "goneFor" was > 0...
this was due to poor if statement placement.

Encapsulated more stuff in exist in try: except:'s and captured errors
outside of it. I believe this will assist with some random exceptions
being thrown when an element didn't exist, when I was testing for
existance.