Added FATAL log type which tells pool to kill us
Currently we aren't properly killing off old phantomjs processes when we
terminate our children, I haven't found a good cause for this yet.

There were errors caused by terminating from a different thread than
the thread that initially spawned that thread, that's resolved mostly.

Fixed log messages a bit so it'll print regardless of if it's a FATAL
error or not