Implemented pause and quit/start, both are =
Added pause and unpause, which for now simply kills the pool. In the
future this'll ideally not send a kill command to a central data
repository.

Quit does the same thing, except the option to start appears after using
it and it's "s - Start". Start simply uses a new option in Pool to go
through and restart the children. Currently this is bugged for Pause and
Quit where the driver instance isn't killed. May be more feasible for
now to simply find a way to halt the child but keep the driver instance
setup.

The menu now contains toggleable options, just for tw of them per entry
though. Hitting p changes the menu option to unpause and then back, this
is all handled in sleep still, which is getting quite crowded.